+++ lib/bencher_api_tests/src/helpers.rs
@@ -1,0 +1,134 @@
+//! Shared test helpers for direct database manipulation.
+//!
+//! These helpers are used by both `api_projects` and `api_runners` integration tests.
+
+use bencher_json::{
+    BranchUuid, DateTime, HeadUuid, JobStatus, JobUuid, ReportUuid, TestbedUuid, VersionUuid,
+};
+use bencher_schema::schema;
+use diesel::{ExpressionMethods as _, QueryDsl as _, RunQueryDsl as _};
+
+use crate::TestServer;
+
+/// Fixed base timestamp for deterministic tests (Unix epoch + 1 billion seconds).
+#[expect(clippy::expect_used)]
+pub fn base_timestamp() -> DateTime {
+    DateTime::try_from(1_000_000_000i64).expect("valid timestamp")
+}
+
+/// Get `project_id` from project slug.
+#[expect(clippy::expect_used)]
+pub fn get_project_id(server: &TestServer, project_slug: &str) -> i32 {
+    let mut conn = server.db_conn();
+    schema::project::table
+        .filter(schema::project::slug.eq(project_slug))
+        .select(schema::project::id)
+        .first(&mut conn)
+        .expect("Failed to get project ID")
+}
+
+/// Create minimal test infrastructure (testbed, version, branch, head, report).
+/// Returns the report ID. Uses a deterministic timestamp.
+#[expect(clippy::expect_used)]
+pub fn create_test_report(server: &TestServer, project_id: i32) -> i32 {
+    let mut conn = server.db_conn();
+    let now = base_timestamp();
+
+    let testbed_uuid = TestbedUuid::new();
+    diesel::insert_into(schema::testbed::table)
+        .values((
+            schema::testbed::uuid.eq(&testbed_uuid),
+            schema::testbed::project_id.eq(project_id),
+            schema::testbed::name.eq("test-testbed"),
+            schema::testbed::slug.eq(&format!("test-testbed-{testbed_uuid}")),
+            schema::testbed::created.eq(&now),
+            schema::testbed::modified.eq(&now),
+        ))
+        .execute(&mut conn)
+        .expect("Failed to insert testbed");
+    let testbed_id: i32 = schema::testbed::table
+        .filter(schema::testbed::uuid.eq(&testbed_uuid))
+        .select(schema::testbed::id)
+        .first(&mut conn)
+        .expect("Failed to get testbed ID");
+
+    let version_uuid = VersionUuid::new();
+    diesel::insert_into(schema::version::table)
+        .values((
+            schema::version::uuid.eq(&version_uuid),
+            schema::version::project_id.eq(project_id),
+            schema::version::number.eq(1),
+        ))
+        .execute(&mut conn)
+        .expect("Failed to insert version");
+    let version_id: i32 = schema::version::table
+        .filter(schema::version::uuid.eq(&version_uuid))
+        .select(schema::version::id)
+        .first(&mut conn)
+        .expect("Failed to get version ID");
+
+    let branch_uuid = BranchUuid::new();
+    diesel::insert_into(schema::branch::table)
+        .values((
+            schema::branch::uuid.eq(&branch_uuid),
+            schema::branch::project_id.eq(project_id),
+            schema::branch::name.eq("main"),
+            schema::branch::slug.eq(&format!("main-{branch_uuid}")),
+            schema::branch::created.eq(&now),
+            schema::branch::modified.eq(&now),
+        ))
+        .execute(&mut conn)
+        .expect("Failed to insert branch");
+    let branch_id: i32 = schema::branch::table
+        .filter(schema::branch::uuid.eq(&branch_uuid))
+        .select(schema::branch::id)
+        .first(&mut conn)
+        .expect("Failed to get branch ID");
+
+    let head_uuid = HeadUuid::new();
+    diesel::insert_into(schema::head::table)
+        .values((
+            schema::head::uuid.eq(&head_uuid),
+            schema::head::branch_id.eq(branch_id),
+            schema::head::created.eq(&now),
+        ))
+        .execute(&mut conn)
+        .expect("Failed to insert head");
+    let head_id: i32 = schema::head::table
+        .filter(schema::head::uuid.eq(&head_uuid))
+        .select(schema::head::id)
+        .first(&mut conn)
+        .expect("Failed to get head ID");
+
+    let report_uuid = ReportUuid::new();
+    diesel::insert_into(schema::report::table)
+        .values((
+            schema::report::uuid.eq(&report_uuid),
+            schema::report::project_id.eq(project_id),
+            schema::report::head_id.eq(head_id),
+            schema::report::version_id.eq(version_id),
+            schema::report::testbed_id.eq(testbed_id),
+            schema::report::adapter.eq(0),
+            schema::report::start_time.eq(&now),
+            schema::report::end_time.eq(&now),
+            schema::report::created.eq(&now),
+        ))
+        .execute(&mut conn)
+        .expect("Failed to insert report");
+
+    schema::report::table
+        .filter(schema::report::uuid.eq(&report_uuid))
+        .select(schema::report::id)
+        .first(&mut conn)
+        .expect("Failed to get report ID")
+}
+
+/// Set job status directly in the database.
+#[expect(clippy::expect_used)]
+pub fn set_job_status(server: &TestServer, job_uuid: JobUuid, status: JobStatus) {
+    let mut conn = server.db_conn();
+    diesel::update(schema::job::table.filter(schema::job::uuid.eq(job_uuid)))
+        .set(schema::job::status.eq(status))
+        .execute(&mut conn)
+        .expect("Failed to set job status");
+}
